[Feat] Added interview details view
Added a view that displays the details of an interview. Changed the interview list to display less information and added a button to see the details of a specific interview. Added a converter class to take responsability away from the interview service. Added a method to the interview service that obtains a single interview.

ID: SM.8